# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/legality/hki/{id}/advance": {
      "patch": {
        "operationId": "LegalityController_advanceHki",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
    "/legality/bpom/{id}/advance": {
      "patch": {
        "operationId": "LegalityController_advanceBpom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
    "/legality/{id}/logs": {
      "get": {
        "operationId": "LegalityController_getLogs",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
...
    "/legality/formula/{id}/validate": {
      "get": {
        "operationId": "LegalityController_validateFormula",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
    "/legality/formula/{id}/review": {
      "post": {
        "operationId": "LegalityController_reviewFormula",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
    "/legality/check-scm/{leadId}": {
      "get": {
        "operationId": "LegalityController_checkScmGate",
        "parameters": [
          {
            "name": "leadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
    "/legality/check-production/{leadId}": {
      "get": {
        "operationId": "LegalityController_checkProductionGate",
        "parameters": [
          {
            "name": "leadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
...
    "/legality/pipeline": {
      "get": {
        "operationId": "LegalityController_getAllPipelines",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "client",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
    "/legality/pipeline/{id}": {
      "get": {
        "operationId": "LegalityController_getPipelineDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Legality"
        ]
      },
      "patch": {
        "operationId": "LegalityController_updatePipeline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
...
    "/legality/pipeline/{id}/artwork-review": {
      "post": {
        "operationId": "LegalityController_submitArtworkReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
    "/legality/pipeline/{id}/pnbp-request": {
      "post": {
        "operationId": "LegalityController_requestPNBP",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
    "/legality/pipeline/{id}/pnbp-pay": {
      "post": {
        "operationId": "LegalityController_payPnbp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
...
    "/legality/master-inci": {
      "get": {
        "operationId": "LegalityController_getMasterIncis",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      },
      "post": {
        "operationId": "LegalityController_createMasterInci",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
    "/legality/master-inci/{id}": {
      "patch": {
        "operationId": "LegalityController_updateMasterInci",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      },
      "delete": {
        "operationId": "LegalityController_deleteMasterInci",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Legality"
        ]
      }
    },
...
    "/fulfillment/shipments": {
      "post": {
        "operationId": "ShipmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShipmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Shipments"
        ]
      },
      "get": {
        "operationId": "ShipmentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Shipments"
        ]
      }
    },
    "/fulfillment/shipments/{id}/status": {
      "patch": {
        "operationId": "ShipmentsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShipmentStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Shipments"
        ]
      }
    },
...
  },
  "info": {
    "title": "NexERP API",
    "description": "The ultimate manufacturing ERP system API documentation",
    "version": "4.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateShipmentDto": {
        "type": "object",
        "properties": {
          "soId": {
            "type": "string"
          },
          "logisticsId": {
            "type": "string",
            "format": "uuid"
          },
          "trackingNo": {
            "type": "string"
          }
        },
        "required": [
          "soId",
          "logisticsId"
        ]
      },
      "UpdateShipmentStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "object"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}